added basic fucnctionality
when the cog is loaded then it will ad one command which will send a ui with 5 options of the main catagories, of problems, then send a url to the md file with the solution.
I plan to automate this with a md parser and search engine, this is only a temporary solution.